perf tools: Add more weak libbpf functions
We hit the window where perf uses libbpf functions, that did not
make it to the official libbpf release yet and it's breaking perf
build with dynamicly linked libbpf.

Fixing this by providing the new interface as weak functions which
calls the original libbpf functions. Fortunatelly the changes were
just renames.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>