static_call: rename EXPORT_ macros to be more self-explanatory
There are currently two ways to export static calls to modules:
- export only the current static call target, so that modules can invoke
  it but not make it point to another target;
- export the static call key so that modules may call
  static_call_update() and patch all invocations tied to that key.

Depending on the flavor of static call implementation, different things
need to be exported: for the generic case, we need to export the key
regardless of which export was chosen, but for other cases, we only
export the trampoline but not the key.

As the latter is an implementation detail, EXPORT_STATIC_CALL_TRAMP() is
not the best choice for the name of the export macro that only exports
the target but not the key.

So let's rename these macros to

  EXPORT_STATIC_CALL        -> EXPORT_STATIC_CALL_FOR_UPDATE
  EXPORT_STATIC_CALL_TRAMP  -> EXPORT_STATIC_CALL

(and the _GPL suffixed ones in a similar fashion). That way, the default
does not expose the key, and users that need to expose control over the
static call key to modules can do so explicitly by choosing the
_FOR_UPDATE version instead.

While at it, add the missing ones for the generic static call
implementation.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>