ceph: don't check for quotas on MDS stray dirs
玮文 胡 reported seeing the WARN_RATELIMITED pop when writing to an
inode that had been transplanted into the stray dir. The client was
trying to look up the quotarealm info from the parent and that tripped
the warning.

Change the ceph_vino_is_reserved helper to not throw a warning and
add a new ceph_vino_warn_reserved() helper that does. Change all of the
existing callsites to call the "warn" variant, and have
ceph_has_realms_with_quotas check return false when the vino is
reserved.

URL: https://tracker.ceph.com/issues/53180
Signed-off-by: Jeff Layton <jlayton@kernel.org>