MAINTAINERS: update information for nouveau
Some side notes on this. Atm we do want to use gitlab for bug tracking and
merge requests. But due to the nature of the current linux kernel
development, we can only do so for nouveau internal changes.

Everything else still needs to be sent as emails and this is also includes
changes to UAPI etc.

Anyway, if somebody wants to submit patches via gitlab, they are free to
do so and this should just make this more official and documented.

People listed as maintainers are such that have push access to drm-misc
(where changes are pushed to after landing in gitlab) and are known
nouveau developers.
We did this already for some trivial changes and critical bug fixes
already, we just weren't thinking about updating the MAINTAINERS file.

Cc: Ben Skeggs <bskeggs@redhat.com>
Cc: Lyude Paul <lyude@redhat.com>
Cc: David Airlie <airlied@linux.ie>
Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: dri-devel@lists.freedesktop.org
Cc: nouveau@lists.freedesktop.org
Signed-off-by: Karol Herbst <kherbst@redhat.com>