KVM: x86/mmu: Factor out part of vmx_get_mt_mask which does not depend on vcpu
Factor out the parts of vmx_get_mt_mask which do not depend on the vCPU
argument. This also requires adding some error reporting to the helper
function to say whether it was possible to generate the MT mask without
a vCPU argument. This refactoring will allow the MT mask to be computed
when noncoherent DMA is not enabled on a VM.

No functional change intended.

Signed-off-by: Ben Gardon <bgardon@google.com>