m68k: Enable memtest kernel parameter
Enable the memtest feature and rearrange some code to prevent it from
clobbering the initrd.

The CONFIG_BLK_DEV_INITRD symbol was conditional on !defined(CONFIG_SUN3).
For simplicity, remove that test on the basis that m68k_ramdisk.size == 0
on Sun 3.

The SLIME source code at http://sammy.net/ shows that no BI_RAMDISK entry
is ever passed to the kernel due to #ifdef 0 around the relevant code.

Cc: Mike Rapoport <rppt@kernel.org>
Cc: Sam Creasey <sammy@sammy.net>
Signed-off-by: Finn Thain <fthain@linux-m68k.org>